Edit on navbar component by adding cart icon and wishlist icon

+++ Backend/src/main/java/com/backend/Service/WishlistService.java
@@ -3,8 +3,8 @@ package com.backend.Service;
 import com.backend.DTO.WishlistDTO;
 
 public interface WishlistService {
-    WishlistDTO getWishlistByUserId(Long userId);
-    WishlistDTO addPieceToWishlist(Long userId, Long pieceId);
-    WishlistDTO removePieceFromWishlist(Long userId, Long pieceId);
-    void clearWishlist(Long userId);
+    WishlistDTO getWishlistByUserId(int userId);
+    WishlistDTO addPieceToWishlist(int userId, int pieceId);
+    WishlistDTO removePieceFromWishlist(int userId, int pieceId);
+    void clearWishlist(int userId);
 }
